requirements.txt
fastapi==0.115.5
Pillow==11.0.0
pydantic==2.10.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
torchvision==0.18.1+cpu
uvicorn==0.32.1
python-multipart==0.0.17
openai==1.12.0
